requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
config==0.5.1
python-decouple
deprecation==2.1.0
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
frozenlist==1.4.1
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kombu==5.4.2
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
openai==1.45.0
packaging==24.1
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
postgrest==0.16.11
prompt_toolkit==3.0.47
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.2
redis==5.0.8
requests==2.32.3
scipy==1.14.1
scrapingbee==2.0.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
storage3==0.7.7
StrEnum==0.4.15
stripe
supabase==2.7.4
supafunc==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.11.1
flower
authlib>=1.0
python-dateutil
openai
python-docx
email_validator
